Playwright downloader init
